de MoreMods en master branch zijn nu samen gevoegd. (de NV registers lopen nu hexadecimaal op, de key counter is nu beveiligd tegen uitvallen en de bugs zijn nu weg gehaald)

+++ ECC_on_wisp/ECC_multiplication/ECC.h
@@ -10,8 +10,8 @@
 
 uint8_t* ECC_multiplication(uint8_t P[], uint8_t n, uint8_t param[]);
 uint8_t* ECC_addition(uint8_t P1[], uint8_t P2[], uint8_t param[]);
-uint8_t ModInv(uint8_t a, uint8_t m);
-uint8_t gcdExtended(uint8_t a, uint8_t b, uint8_t *x, uint8_t *y);
+uint16_t ModInv(uint16_t a, uint16_t m);
+uint16_t gcdExtended(uint16_t a, uint16_t b, uint16_t *x, uint16_t *y);
 uint8_t ipow(uint8_t base, uint8_t exp, uint8_t m);
 void main_ecc(uint8_t param[], uint8_t start[], uint8_t privateKey);
